+++ src/Microsoft.ML.CpuMath/CpuMathUtils.DotNetCoreApp.cs
@@ -10,34 +10,46 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 {
     public static partial class CpuMathUtils
     {
-        private static unsafe Vector128<float> Load1(float* src, int* idx)
+        public const int CbAlign = 16;
+
+        private static bool Compat(AlignedArray a)
         {
-            return Sse.SetScalarVector128(src[idx[0]]);
+            Contracts.AssertValue(a);
+            Contracts.Assert(a.Size > 0);
+            return a.CbAlign == CbAlign;
         }
 
-        private static unsafe Vector128<float> Load4(float* src, int* idx)
+        private unsafe static float* Ptr(AlignedArray a, float* p)
         {
-            return Sse.SetVector128(src[idx[3]], src[idx[2]], src[idx[1]], src[idx[0]]);
+            Contracts.AssertValue(a);
+            float* q = p + a.GetBase((long)p);
+            Contracts.Assert(((long)q & (CbAlign - 1)) == 0);
+            return q;
         }
 
-        private static Vector128<float> Rotate(Vector128<float> x)
+        public static void MatTimesSrc(bool tran, bool add, AlignedArray mat, AlignedArray src, AlignedArray dst, int crun)
         {
-            return Sse.Shuffle(x, x, 0x39);
-        }
+            Contracts.Assert(Compat(mat));
+            Contracts.Assert(Compat(src));
+            Contracts.Assert(Compat(dst));
+            Contracts.Assert(mat.Size == dst.Size * src.Size);
 
-        private static Vector128<float> RotateReverse(Vector128<float> x)
-        {
-            return Sse.Shuffle(x, x, 0x93);
-        }
-
-        // Warning: this operation changes the value of x => do not reuse x
-        private static unsafe void Store4(Vector128<float> x, float* dst, int* idx)
-        {
-            Sse.StoreScalar(dst + idx[0], x);
-            for (int i = 1; i <= 3; i++)
+            if (Sse.IsSupported)
             {
-                x = Rotate(x);
-                Sse.StoreScalar(dst + idx[i], x);
+                if (!tran)
+                {
+                    Contracts.Assert(0 <= crun && crun <= dst.Size);
+                    SseIntrinsics.MatMulASse(add, mat, src, dst, crun, src.Size);
+                }
+                else
+                {
+                    Contracts.Assert(0 <= crun && crun <= src.Size);
+                    SseIntrinsics.MatMulTranA(add, mat, src, dst, dst.Size, crun);
+                }
+            }
+            else
+            {
+                // TODO: Software fallback.
             }
         }
 
@@ -48,47 +60,13 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                ScaleUSse(a, dst, count);
+                SseIntrinsics.ScaleU(a, dst, count);
             }
             else
             {
                 for (int i = 0; i < dst.Length; i++)
                 {
                     dst[i] *= a;
-                }
-            }
-        }
-
-        private static void ScaleUSse(float scale, float[] dst, int count)
-        {
-            Vector128<float> scaleVector = Sse.SetAllVector128(scale);
-
-            unsafe
-            {
-                fixed (float* pdst = dst)
-                {
-                    float* pDstCurrent = pdst;
-                    float* pEnd = pdst + count;
-
-                    while (pDstCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> dstVector = Sse.LoadVector128(pDstCurrent);
-
-                        dstVector = Sse.Multiply(scaleVector, dstVector);
-                        Sse.Store(pDstCurrent, dstVector);
-
-                        pDstCurrent += 4;
-                    }
-
-                    while (pDstCurrent < pEnd)
-                    {
-                        Vector128<float> dstVector = Sse.LoadScalarVector128(pDstCurrent);
-
-                        dstVector = Sse.MultiplyScalar(scaleVector, dstVector);
-                        Sse.StoreScalar(pDstCurrent, dstVector);
-
-                        pDstCurrent++;
-                    }
                 }
             }
         }
@@ -112,55 +90,13 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                AddScaleUSse(a, src, dst, count);
+                SseIntrinsics.AddScaleU(a, src, dst, count);
             }
             else
             {
                 for (int i = 0; i < count; i++)
                 {
                     dst[i] += a * src[i];
-                }
-            }
-        }
-
-        private static void AddScaleUSse(float scale, float[] src, float[] dst, int count)
-        {
-            Vector128<float> scaleVector = Sse.SetAllVector128(scale);
-
-            unsafe
-            {
-                fixed (float* psrc = src)
-                fixed (float* pdst = dst)
-                {
-                    float* pSrcCurrent = psrc;
-                    float* pDstCurrent = pdst;
-                    float* pEnd = pdst + count;
-
-                    while (pDstCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadVector128(pSrcCurrent);
-                        Vector128<float> dstVector = Sse.LoadVector128(pDstCurrent);
-
-                        srcVector = Sse.Multiply(srcVector, scaleVector);
-                        dstVector = Sse.Add(dstVector, srcVector);
-                        Sse.Store(pDstCurrent, dstVector);
-
-                        pDstCurrent += 4;
-                        pSrcCurrent += 4;
-                    }
-
-                    while (pDstCurrent < pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadScalarVector128(pSrcCurrent);
-                        Vector128<float> dstVector = Sse.LoadScalarVector128(pDstCurrent);
-
-                        srcVector = Sse.MultiplyScalar(srcVector, scaleVector);
-                        dstVector = Sse.AddScalar(dstVector, srcVector);
-                        Sse.StoreScalar(pDstCurrent, dstVector);
-
-                        pDstCurrent++;
-                        pSrcCurrent++;
-                    }
                 }
             }
         }
@@ -188,7 +124,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                AddScaleSUSse(a, src, indices, dst, count);
+                SseIntrinsics.AddScaleSU(a, src, indices, dst, count);
             }
             else
             {
@@ -197,46 +133,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                     dst[indices[i]] += a * src[i];
                 }
             }
-        }
-
-        private static void AddScaleSUSse(float scale, float[] src, int[] idx, float[] dst, int count)
-        {
-            Vector128<float> scaleVector = Sse.SetAllVector128(scale);
-
-            unsafe
-            {
-                fixed (float* psrc = src)
-                fixed (int* pidx = idx)
-                fixed (float* pdst = dst)
-                {
-                    float* pSrcCurrent = psrc;
-                    int* pIdxCurrent = pidx;
-                    float* pDstCurrent = pdst;
-                    int* pEnd = pidx + count;
-
-                    while (pIdxCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadVector128(pSrcCurrent);
-                        Vector128<float> dstVector = Load4(pDstCurrent, pIdxCurrent);
-
-                        srcVector = Sse.Multiply(srcVector, scaleVector);
-                        dstVector = Sse.Add(dstVector, srcVector);
-                        Store4(dstVector, pDstCurrent, pIdxCurrent);
-
-                        pIdxCurrent += 4;
-                        pSrcCurrent += 4;
-                    }
-
-                    while (pIdxCurrent < pEnd)
-                    {
-                        pDstCurrent[*pIdxCurrent] += scale * *pSrcCurrent;
-
-                        pIdxCurrent++;
-                        pSrcCurrent++;
-                    }
-                }
-            }
-        }
+        } 
 
         public static void AddScale(float a, float[] src, int[] indices, float[] dst, int dstOffset, int count)
         {
@@ -261,51 +158,13 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                AddUSse(src, dst, count);
+                SseIntrinsics.AddU(src, dst, count);
             }
             else
             {
                 for (int i = 0; i < count; i++)
                 {
                     dst[i] += src[i];
-                }
-            }
-        }
-
-        private static void AddUSse(float[] src, float[] dst, int count)
-        {
-            unsafe
-            {
-                fixed (float* psrc = src)
-                fixed (float* pdst = dst)
-                {
-                    float* pSrcCurrent = psrc;
-                    float* pDstCurrent = pdst;
-                    float* pEnd = psrc + count;
-
-                    while (pSrcCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadVector128(pSrcCurrent);
-                        Vector128<float> dstVector = Sse.LoadVector128(pDstCurrent);
-
-                        Vector128<float> result = Sse.Add(srcVector, dstVector);
-                        Sse.Store(pDstCurrent, result);
-
-                        pSrcCurrent += 4;
-                        pDstCurrent += 4;
-                    }
-
-                    while (pSrcCurrent < pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadScalarVector128(pSrcCurrent);
-                        Vector128<float> dstVector = Sse.LoadScalarVector128(pDstCurrent);
-
-                        Vector128<float> result = Sse.AddScalar(srcVector, dstVector);
-                        Sse.StoreScalar(pDstCurrent, result);
-
-                        pSrcCurrent++;
-                        pDstCurrent++;
-                    }
                 }
             }
         }
@@ -321,49 +180,13 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                AddSUSse(src, indices, dst, count);
+                SseIntrinsics.AddSU(src, indices, dst, count);
             }
             else
             {
                 for (int i = 0; i < count; i++)
                 {
                     dst[indices[i]] += src[i];
-                }
-            }
-        }
-
-        private static void AddSUSse(float[] src, int[] idx, float[] dst, int count)
-        {
-            unsafe
-            {
-                fixed (float* psrc = src)
-                fixed (int* pidx = idx)
-                fixed (float* pdst = dst)
-                {
-                    float* pSrcCurrent = psrc;
-                    int* pIdxCurrent = pidx;
-                    float* pDstCurrent = pdst;
-                    int* pEnd = pidx + count;
-
-                    while (pIdxCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> srcVector = Load4(pDstCurrent, pIdxCurrent);
-                        Vector128<float> dstVector = Sse.LoadVector128(pSrcCurrent);
-
-                        srcVector = Sse.Add(srcVector, dstVector);
-                        Store4(srcVector, pDstCurrent, pIdxCurrent);
-
-                        pIdxCurrent += 4;
-                        pSrcCurrent += 4;
-                    }
-
-                    while (pIdxCurrent < pEnd)
-                    {
-                        pDstCurrent[*pIdxCurrent] += *pSrcCurrent;
-
-                        pIdxCurrent++;
-                        pSrcCurrent++;
-                    }
                 }
             }
         }
@@ -392,53 +215,13 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                MulElementWiseUSse(src1, src2, dst, count);
+                SseIntrinsics.MulElementWiseU(src1, src2, dst, count);
             }
             else
             {
                 for (int i = 0; i < count; i++)
                 {
                     dst[i] = src1[i] * src2[i];
-                }
-            }
-        }
-
-        private static void MulElementWiseUSse(float[] src1, float[] src2, float[] dst, int count)
-        {
-            unsafe
-            {
-                fixed (float* psrc1 = &src1[0])
-                fixed (float* psrc2 = &src2[0])
-                fixed (float* pdst = dst)
-                {
-                    float* pSrc1Current = psrc1;
-                    float* pSrc2Current = psrc2;
-                    float* pDstCurrent = pdst;
-                    float* pEnd = pdst + count;
-
-                    while (pDstCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> src1Vector = Sse.LoadVector128(pSrc1Current);
-                        Vector128<float> src2Vector = Sse.LoadVector128(pSrc2Current);
-                        src2Vector = Sse.Multiply(src1Vector, src2Vector);
-                        Sse.Store(pDstCurrent, src2Vector);
-
-                        pSrc1Current += 4;
-                        pSrc2Current += 4;
-                        pDstCurrent += 4;
-                    }
-
-                    while (pDstCurrent < pEnd)
-                    {
-                        Vector128<float> src1Vector = Sse.LoadScalarVector128(pSrc1Current);
-                        Vector128<float> src2Vector = Sse.LoadScalarVector128(pSrc2Current);
-                        src2Vector = Sse.MultiplyScalar(src1Vector, src2Vector);
-                        Sse.StoreScalar(pDstCurrent, src2Vector);
-
-                        pSrc1Current++;
-                        pSrc2Current++;
-                        pDstCurrent++;
-                    }
                 }
             }
         }
@@ -450,7 +233,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                return SumSqUSse(src, count);
+                return SseIntrinsics.SumSqU(src, count);
             }
             else
             {
@@ -462,51 +245,6 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                 }
                 return result;
             }
-        }
-
-        private static float SumSqUSse(float[] src, int count)
-        {
-            Vector128<float> result = Sse.SetZeroVector128();
-
-            unsafe
-            {
-                fixed (float* psrc = src)
-                {
-                    float* pSrcCurrent = psrc;
-                    float* pEnd = psrc + count;
-
-                    while (pSrcCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadVector128(pSrcCurrent);
-                        result = Sse.Add(result, Sse.Multiply(srcVector, srcVector));
-
-                        pSrcCurrent += 4;
-                    }
-
-                    if (Sse3.IsSupported)
-                    {
-                        // SSE3 is supported.
-                        result = Sse3.HorizontalAdd(result, result);
-                        result = Sse3.HorizontalAdd(result, result);
-                    }
-                    else
-                    {
-                        // SSE3 is not supported.
-                        result = Sse.Add(result, Sse.MoveHighToLow(result, result));
-                        result = Sse.Add(result, Sse.MoveHighToLow(result, Sse.UnpackLow(result, result)));
-                    }
-
-                    while (pSrcCurrent < pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadScalarVector128(pSrcCurrent);
-                        result = Sse.AddScalar(result, Sse.MultiplyScalar(srcVector, srcVector));
-
-                        pSrcCurrent++;
-                    }
-                }
-            }
-
-            return Sse.ConvertToSingle(result);
         }
 
         public static float SumSq(float[] src, int offset, int count)
@@ -526,7 +264,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                return SumAbsUSse(src, count);
+                return SseIntrinsics.SumAbsU(src, count);
             }
             else
             {
@@ -537,62 +275,6 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                 }
                 return sum;
             }
-        }
-
-        private static float SumAbsUSse(float[] src, int count)
-        {
-            Vector128<float> result = Sse.SetZeroVector128();
-            Vector128<float> mask;
-
-            if (Sse2.IsSupported)
-            {
-                mask = Sse.StaticCast<int, float>(Sse2.SetAllVector128(0x7FFFFFFF));
-            }
-            else
-            {
-                mask = Sse.SetAllVector128(0x7FFFFFFF);
-            }
-
-            unsafe
-            {
-                fixed (float* psrc = src)
-                {
-                    float* pSrcCurrent = psrc;
-                    float* pEnd = psrc + count;
-
-                    while (pSrcCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadVector128(pSrcCurrent);
-                        result = Sse.Add(result, Sse.And(srcVector, mask));
-
-                        pSrcCurrent += 4;
-                    }
-
-                    // TODO: Check code length per line
-                    if (Sse3.IsSupported)
-                    {
-                        // SSE3 is supported.
-                        result = Sse3.HorizontalAdd(result, result);
-                        result = Sse3.HorizontalAdd(result, result);
-                    }
-                    else
-                    {
-                        // SSE3 is not supported.
-                        result = Sse.Add(result, Sse.MoveHighToLow(result, result));
-                        result = Sse.Add(result, Sse.MoveHighToLow(result, Sse.UnpackLow(result, result)));
-                    }
-
-                    while (pSrcCurrent < pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadVector128(pSrcCurrent);
-                        result = Sse.Add(result, Sse.And(srcVector, mask));
-
-                        pSrcCurrent++;
-                    }
-                }
-            }
-
-            return Sse.ConvertToSingle(result);
         }
 
         public static float SumAbs(float[] src, int offset, int count)
@@ -615,7 +297,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                return DotUSse(a, b, count);
+                return SseIntrinsics.DotU(a, b, count);
             }
             else
             {
@@ -627,61 +309,6 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                 }
                 return result;
             }
-        }
-
-        private static float DotUSse(float[] src, float[] dst, int count)
-        {
-            Vector128<float> result = Sse.SetZeroVector128();
-
-            unsafe
-            {
-
-                fixed (float* psrc = src)
-                fixed (float* pdst = dst)
-                {
-                    float* pSrcCurrent = psrc;
-                    float* pDstCurrent = pdst;
-                    float* pEnd = psrc + count;
-
-                    while (pSrcCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadVector128(pSrcCurrent);
-                        Vector128<float> dstVector = Sse.LoadVector128(pDstCurrent);
-
-                        result = Sse.Add(result, Sse.Multiply(srcVector, dstVector));
-
-                        pSrcCurrent += 4;
-                        pDstCurrent += 4;
-                    }
-
-                    if (Sse3.IsSupported)
-                    {
-                        // SSE3 is supported.
-                        result = Sse3.HorizontalAdd(result, result);
-                        result = Sse3.HorizontalAdd(result, result);
-                    }
-                    else
-                    {
-                        // SSE3 is not supported.
-                        result = Sse.Add(result, Sse.MoveHighToLow(result, result));
-                        result = Sse.Add(result, Sse.MoveHighToLow(result, Sse.UnpackLow(result, result)));
-                    }
-
-
-                    while (pSrcCurrent < pEnd)
-                    {
-                        Vector128<float> srcVector = Sse.LoadScalarVector128(pSrcCurrent);
-                        Vector128<float> dstVector = Sse.LoadScalarVector128(pDstCurrent);
-
-                        result = Sse.AddScalar(result, Sse.MultiplyScalar(srcVector, dstVector));
-
-                        pSrcCurrent++;
-                        pDstCurrent++;
-                    }
-                }
-            }
-
-            return Sse.ConvertToSingle(result);
         }
 
         public static float DotProductDense(float[] a, int offset, float[] b, int count)
@@ -707,7 +334,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                return DotSUSse(a, b, indices, count);
+                return SseIntrinsics.DotSU(a, b, indices, count);
             }
             else
             {
@@ -719,62 +346,6 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                 }
                 return result;
             }
-        }
-
-        private static float DotSUSse(float[] src, float[] dst, int[] idx, int count)
-        {
-            Vector128<float> result = Sse.SetZeroVector128();
-
-            unsafe
-            {
-                fixed (float* psrc = src)
-                fixed (float* pdst = dst)
-                fixed (int* pidx = idx)
-                {
-                    float* pSrcCurrent = psrc;
-                    float* pDstCurrent = pdst;
-                    int* pIdxCurrent = pidx;
-                    int* pEnd = pidx + count;
-
-                    while (pIdxCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> srcVector = Load4(pSrcCurrent, pIdxCurrent);
-                        Vector128<float> dstVector = Sse.LoadVector128(pDstCurrent);
-
-                        result = Sse.Add(result, Sse.Multiply(srcVector, dstVector));
-
-                        pIdxCurrent += 4;
-                        pDstCurrent += 4;
-                    }
-
-                    if (Sse3.IsSupported)
-                    {
-                        // SSE3 is supported.
-                        result = Sse3.HorizontalAdd(result, result);
-                        result = Sse3.HorizontalAdd(result, result);
-                    }
-                    else
-                    {
-                        // SSE3 is not supported.
-                        result = Sse.Add(result, Sse.MoveHighToLow(result, result));
-                        result = Sse.Add(result, Sse.MoveHighToLow(result, Sse.UnpackLow(result, result)));
-                    }
-
-
-                    while (pIdxCurrent < pEnd)
-                    {
-                        Vector128<float> srcVector = Load1(pSrcCurrent, pIdxCurrent);
-                        Vector128<float> dstVector = Sse.LoadScalarVector128(pDstCurrent);
-
-                        result = Sse.AddScalar(result, Sse.MultiplyScalar(srcVector, dstVector));
-
-                        pIdxCurrent++;
-                        pDstCurrent++;
-                    }
-                }
-            }
-
-            return Sse.ConvertToSingle(result);
         }
 
         public static float DotProductSparse(float[] a, int offset, float[] b, int[] indices, int count)
@@ -800,7 +371,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
             if (Sse.IsSupported)
             {
-                return Dist2Sse(a, b, count);
+                return SseIntrinsics.Dist2(a, b, count);
             }
             else
             {
@@ -814,55 +385,5 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        private static float Dist2Sse(float[] src, float[] dst, int count)
-        {
-            Vector128<float> SqDistanceVector = Sse.SetZeroVector128();
-
-            unsafe
-            {
-                fixed (float* psrc = src)
-                fixed (float* pdst = dst)
-                {
-                    float* pSrcCurrent = psrc;
-                    float* pDstCurrent = pdst;
-                    float* pEnd = psrc + count;
-
-                    while (pSrcCurrent + 4 <= pEnd)
-                    {
-                        Vector128<float> distanceVector = Sse.Subtract(Sse.LoadVector128(pSrcCurrent), Sse.LoadVector128(pDstCurrent));
-                        SqDistanceVector = Sse.Add(SqDistanceVector, Sse.Multiply(distanceVector, distanceVector));
-
-                        pSrcCurrent += 4;
-                        pDstCurrent += 4;
-                    }
-
-                    // TODO: Check code length per line
-                    if (Sse3.IsSupported)
-                    {
-                        // SSE3 is supported.
-                        SqDistanceVector = Sse3.HorizontalAdd(SqDistanceVector, SqDistanceVector);
-                        SqDistanceVector = Sse3.HorizontalAdd(SqDistanceVector, SqDistanceVector);
-                    }
-                    else
-                    {
-                        // SSE3 is not supported.
-                        SqDistanceVector = Sse.Add(SqDistanceVector, Sse.MoveHighToLow(SqDistanceVector, SqDistanceVector));
-                        SqDistanceVector = Sse.Add(SqDistanceVector, Sse.MoveHighToLow(SqDistanceVector, Sse.UnpackLow(SqDistanceVector, SqDistanceVector)));
-                    }
-
-                    float norm = Sse.ConvertToSingle(SqDistanceVector);
-                    while (pSrcCurrent < pEnd)
-                    {
-                        float distance = *pSrcCurrent - *pDstCurrent;
-                        norm += distance * distance;
-
-                        pSrcCurrent++;
-                        pDstCurrent++;
-                    }
-
-                    return norm;
-                }
-            }
-        }
     }
 }
